requirements.txt
fastapi==0.100.1
fastapi-jwt==0.1.12
uvicorn[standard]
gunicorn
databases==0.7.0
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic[email]==2.1.1
pydantic-settings==2.0.2
passlib[bcrypt]==1.7.4
pyhumps==3.8.0
sentry-sdk[fastapi]==1.29.2
psycopg2-binary==2.9.7
prometheus-fastapi-instrumentator==6.1.0
alembic==1.12.0
pyTelegramBotAPI==4.14.0
celery[redis]==5.3.4
httpx==0.25.0
flower==2.0.1
websockets==11.0.3
validate_email==1.3